requirements.txt
python-telegram-bot[job-queue]>=21.0,<23.0
python-dotenv==1.0.1
